+++ R/ship_of_theseus.R
@@ -5,10 +5,9 @@ NULL
 #'
 #' @description
 #' The \code{ShipOfTheseus} class decomposes the difference in outcome rates
-#' between two datasets and visualizes the results as a Theseus Plot. It
-#' provides methods to compute contributions of individual attributes, summarize
-#' results in tables, and generate waterfall-style plots for intuitive
-#' interpretation.
+#' between two datasets. For a selected column, it computes subgroup
+#' contributions, summarizes the results in tables, and visualizes them as
+#' waterfall-style Theseus Plots.
 #'
 #' @import dplyr ggplot2 stringr
 #'
@@ -23,6 +22,7 @@ ShipOfTheseus <- R6::R6Class(
     compute_contribution = NULL,
     compute_info = NULL,
     compute_size = NULL,
+    xlab = NULL,
     ylab = NULL,
     digits = NULL,
     text_size = NULL,
@@ -35,10 +35,60 @@ ShipOfTheseus <- R6::R6Class(
         # table()/plot() handle character and factor columns differently
         mutate(
           across(where(is.character), \(x) replace_na(x, "(Missing)")),
-          across(where(is.factor), \(x) fct_na_value_to_level(x, level = "(Missing)"))
+          across(where(is.factor) & where(anyNA), \(x) fct_na_value_to_level(x, level = "(Missing)"))
         ) |>
         # Standardize the outcome column name for internal use
         rename(.outcome = all_of(outcome))
+    },
+    
+    prepare_plot_data = function(column_name, n, levels, continuous, flip) {
+      df_contrib <- self$table(!!rlang::sym(column_name), n = n, continuous = continuous) |>
+        select(items = 1L, contrib)
+      
+      if (flip) {
+        df_contrib <- df_contrib |> mutate(contrib = -contrib)
+      }
+      
+      is_factor <- is.factor(df_contrib$items)
+      
+      if (is_factor) {
+        df_contrib <- if (flip) arrange(df_contrib, desc(items)) else arrange(df_contrib, items)
+      } else {
+        df_contrib <- arrange(df_contrib, contrib)
+      }
+      
+      data_size <- private$compute_size(
+        column_name,
+        target = df_contrib$items,
+        continuous = continuous
+      )
+      
+      if (!is.null(levels)) {
+        levels <- as.character(levels)
+        if (flip) levels <- rev(levels)
+        df_contrib <- data.frame(items = levels) |> inner_join(df_contrib, by = "items")
+      }
+      
+      list(
+        df_contrib = df_contrib,
+        data_size = data_size
+      )
+    },
+    
+    compute_bar_height = function(data_waterfall, data_size, max_score, main_item = NULL,
+                                  bar_max_value = NULL, bar_min_ratio = 0.15) {
+      if (is.null(main_item) & is.null(bar_max_value)) {
+        max_amount <- max_score * 100 * bar_min_ratio
+        n_max <- data_size |> tail(-1) |> 
+          slice_max(order_by = n, n = 1, with_ties = FALSE) |> pull(n)
+      } else if(!is.null(main_item)) {
+        max_amount <- data_waterfall |> filter(items == main_item) |> pull(contrib) |> abs()
+        n_max <- data_size |> filter(items == main_item) |> pull(n) |> max()
+      } else if(!is.null(bar_max_value)) {
+        max_amount <- bar_max_value
+        n_max <- data_size |> filter(n == max(n)) |> pull(n) |> max()
+      }
+      max_amount / n_max
     }
   ),
 
@@ -47,21 +97,23 @@ ShipOfTheseus <- R6::R6Class(
     #' The constructor of the \code{ShipOfTheseus} class.
     #'
     #' @param data1 data frame representing the first group (e.g., the baseline
-    #'   or "original" data).
+    #'   data).
     #' @param data2 data frame representing the second group (e.g., the
-    #'   comparison or "refitted" data).
+    #'   comparison data).
     #' @param outcome string specifying the outcome variable used to compute the
     #'   rate metric (default is \code{"y"}). Typically, this is a binary
     #'   indicator (e.g., 0/1) that is aggregated to form rates.
     #' @param labels character vector of length 2 giving the labels for the two
     #'   groups. The first corresponds to \code{data1}, the second to
-    #'   \code{data2}. Default is \code{c("Original", "Refitted")}.
+    #'   \code{data2}. Default is \code{c("Baseline", "Comparison")}.
+    #' @param xlab string specifying the x-axis label for plots. If \code{NULL}
+    #'   (default), no label is displayed.
     #' @param ylab string specifying the y-axis label for plots. If \code{NULL}
     #'   (default), no label is displayed.
     #' @param digits integer indicating the number of decimal places to use for
-    #'   displaying numeric values (default is \code{3}).
+    #'   displaying numeric values (default is \code{1}).
     #' @param text_size numeric value specifying the relative size of text
-    #'   elements in plots (default is \code{1}).
+    #'   elements in plots (default is \code{1.0}).
     #'
     #' @return A \code{ShipOfTheseus} object, which can be used with
     #'   \code{plot()} to create Theseus plots.
@@ -70,13 +122,15 @@ ShipOfTheseus <- R6::R6Class(
     #' @importFrom memoise memoise
     #' @importFrom tibble tibble
     #' @importFrom tidyr replace_na
-    initialize = function(data1, data2, outcome, labels, ylab, digits, text_size) {
+    initialize = function(data1, data2, outcome, labels, xlab, ylab, digits, text_size) {
       outcome <- rlang::quo_squash(outcome) |> rlang::as_string()
+      labels <- as.character(labels)
 
       data1 <- private$prepare_input_data(data1, outcome)
       data2 <- private$prepare_input_data(data2, outcome)
 
       private$labels <- labels
+      private$xlab <- xlab 
       private$ylab <- ylab
       private$digits <- digits
       private$text_size <- text_size
@@ -167,7 +221,7 @@ ShipOfTheseus <- R6::R6Class(
         score1 <- scores[1]
         score2 <- scores[2]
 
-        result <- tibble()
+        df_contrib <- tibble()
         for (name in names2) {
           df_temp <- df1
           if (name %in% names1) {
@@ -179,7 +233,7 @@ ShipOfTheseus <- R6::R6Class(
           score_new <- df_temp |> summarise(score = sum(y) / sum(n)) |> pull(score)
           diff <- score_new - score1
           res <- tibble(items = name, amount = diff)
-          result <- rbind(result, res)
+          df_contrib <- rbind(df_contrib, res)
         }
         for (name in names1) {
           df_temp <- df2
@@ -192,15 +246,15 @@ ShipOfTheseus <- R6::R6Class(
           score_new <- df_temp |> summarise(score = sum(y) / sum(n)) |> pull(score)
           diff <- score2 - score_new
           res <- tibble(items = name, amount = diff)
-          result <- rbind(result, res)
+          df_contrib <- rbind(df_contrib, res)
         }
 
         if (is.factor(names1)) {
           names <- fct_c(names1, names2)
-          result <- result |>
+          df_contrib <- df_contrib |>
             mutate(items = factor(items, levels = levels(names)))
         }
-        result |>
+        df_contrib |>
           group_by(items) |>
           summarise(contrib = mean(amount)) |>
           mutate(contrib = (score2 - score1) * contrib / sum(contrib))
@@ -262,21 +316,20 @@ ShipOfTheseus <- R6::R6Class(
           }
         }
       })
-
     },
 
     #' @description
     #' Generate a contribution table for a given column.
     #'
     #' @param column_name string. The name of the column to analyze.
-    #' @param n integer. Maximum number of top contributing attributes to display.
-    #'   If the number of attributes exceeds `n`, the remaining are aggregated.
+    #' @param n integer. Maximum number of top contributing subgroups to display.
+    #'   If the number of subgroups exceeds `n`, the remaining are aggregated.
     #' @param continuous list. A configuration list for handling continuous
     #'   variables (e.g., specifying number of bins or custom breaks).
     #'
-    #' @return A tibble summarizing each attribute's contribution to the
-    #'   difference between the two groups, including counts, total outcomes,
-    #'   and rates for each subgroup.
+    #' @return A tibble summarizing subgroup contributions to the difference
+    #'   between the two groups, including counts, total outcomes, and rates
+    #'   for each subgroup.
     table = function(column_name, n = Inf, continuous = continuous_config()) {
       column_name <- rlang::ensym(column_name) |> rlang::as_string()
       data_contrib <- private$compute_contribution(column_name, continuous)
@@ -297,13 +350,13 @@ ShipOfTheseus <- R6::R6Class(
           mutate(items = as.character(items))
         result_tail <- tail(result |> arrange(desc(abs(contrib))), n_other)
         result_other <- result_tail |>
-          mutate(items = str_glue("Sum of {n_other} other attributes")) |>
+          mutate(items = str_glue("Sum of {n_other} other subgroups")) |>
           group_by(items) |>
           summarise_at(vars(contrib, n1, n2, x1, x2), sum) |>
           mutate(rate1 = x1 / n1, rate2 = x2 / n2)
         result <- bind_rows(result_head, result_other)
         if (is_factor) {
-          levels <- c(levels, str_glue("Sum of {n_other} other attributes"))
+          levels <- c(levels, str_glue("Sum of {n_other} other subgroups"))
           result <- result |>
             mutate(items = factor(items, levels = levels)) |>
             arrange(items)
@@ -317,18 +370,17 @@ ShipOfTheseus <- R6::R6Class(
     #' Generate a Theseus plot for a specified column
     #'
     #' @param column_name The name of the column to visualize.
-    #' @param n integer. Maximum number of top contributing attributes to display.
-    #'   Remaining attributes are aggregated if necessary.
-    #' @param main_item string. The attribute used as the reference for scaling
+    #' @param n integer. Maximum number of top contributing subgroups to display.
+    #'   Remaining subgroups are aggregated if necessary.
+    #' @param main_item string. The subgroup used as the reference for scaling
     #'   the bar heights.
     #' @param bar_max_value numeric. Maximum value for scaling the contribution bars.
-    #' @param levels character vector specifying the display order of attributes.
+    #' @param levels character vector specifying the display order of subgroups.
     #' @param continuous list. Configuration for handling continuous variables
     #'   (e.g., number of bins or custom breaks).
     #'
     #' @return A ggplot object representing the Theseus Plot for the specified column.
     #'
-    #' @importFrom tidyr replace_na
     #' @importFrom tibble tibble
     #' @importFrom waterfalls waterfall
     plot = function(column_name, n = 10L, main_item = NULL, bar_max_value = NULL,
@@ -336,76 +388,58 @@ ShipOfTheseus <- R6::R6Class(
       column_name <- rlang::ensym(column_name) |> rlang::as_string()
 
       labels <- private$labels
+      scores <- private$compute_scores(column_name)
+      score1 <- scores[1]
+      max_score <- max(scores)
+      plot_data <- private$prepare_plot_data(column_name, n, levels, continuous, flip = FALSE)
+      
+      df_contrib <- plot_data$df_contrib
+      data_size <- plot_data$data_size
 
-      score1 <- private$compute_scores(column_name)[1]
-
-      result <- self$table(!!rlang::sym(column_name), n = n, continuous = continuous) |>
-        select(items = 1L, contrib)
-      is_factor <- is.factor(result$items)
-      if (is_factor) {
-        result <- result |> arrange(items)
-      } else {
-        result <- result |> arrange(contrib)
-      }
-
-      data_size <- private$compute_size(column_name, target = result$items, continuous = continuous)
-
-      if (!is.null(levels)) {
-        levels <- as.character(levels)
-        result <- data.frame(items = levels) |> inner_join(result, by = "items")
-      }
-      names <- as.character(result$items)
-      result <- tibble::tibble(items = labels[1], contrib = score1) |>
-        bind_rows(result)|>
+      names <- as.character(df_contrib$items)
+      data_waterfall <- tibble(items = labels[1], contrib = score1) |>
+        bind_rows(df_contrib)|>
         mutate(contrib = round(contrib * 100, digits = private$digits))
 
-      p <- suppressWarnings({
-        waterfalls::waterfall(
-          result, calc_total = TRUE, total_axis_text = labels[2],
-          total_rect_text_color = "black", total_rect_color = "#00BFC4", rect_text_size = private$text_size)
-      })
+      n_bars <- nrow(data_waterfall) + 1L
+      p <- waterfall(
+        data_waterfall,
+        calc_total = TRUE, total_axis_text = n_bars,
+        total_rect_text_color = "black", total_rect_color = "#00BFC4", 
+        rect_text_size = private$text_size)
+      scale_x <- p@scales$get_scales("x")
+      scale_x$labels <- c(data_waterfall$items, labels[2])
+      
+      bar_rate <- private$compute_bar_height(data_waterfall, data_size, max_score, 
+                                             main_item, bar_max_value)
 
-      if (is.null(main_item) & is.null(bar_max_value)) {
-        data_max <- result |> tail(-1) |> 
-          slice_max(order_by = abs(contrib), n = 1, with_ties = FALSE)
-        max_item <- data_max |> pull(items)
-        max_amount <- data_max |> pull(contrib) |> abs()
-        n_max <- data_size |> filter(items == max_item) |> pull(n) |> max()
-      } else if(!is.null(main_item)) {
-        max_amount <- result |> filter(items == main_item) |> pull(contrib) |> abs()
-        n_max <- data_size |> filter(items == main_item) |> pull(n) |> max()
-      } else if(!is.null(bar_max_value)) {
-        max_amount <- bar_max_value
-        n_max <- data_size |> filter(n == max(n)) |> pull(n) |> max()
-      }
-
-      levels <- c(labels[1], names, labels[2])
-      data_size <- p$data |> select(x) |> distinct() |>
-        left_join(data_size, by = c(x = "items")) |>
+      data_size <- tibble(x = factor(seq_len(n_bars))) |>
+        mutate(items = c(labels[1], names, labels[2])) |>
+        left_join(data_size, by = "items") |>
         tidyr::replace_na(list(n = 0L)) |>
-        mutate(x = factor(x, levels = levels), type = factor(type, levels = labels)) |>
-        mutate(n = n / n_max * max_amount)
-
+        mutate(type = factor(type, levels = labels)) |>
+        mutate(n = n * bar_rate)
+            
       p <- p +
         geom_col(data = data_size, aes(x, n, fill = type), width = 0.7, position = position_dodge()) +
         scale_fill_manual(values = c("#7CAE00", "#C77CFF"), guide = "none")
       p$layers <- append(head(p$layers, -1), tail(p$layers, 1), 1)
-      theme <- ggplot2::theme_get()
-      p + ggplot2::ggtitle(NULL, subtitle = column_name) +
-        theme(private$text_size * 11) +
-        ggplot2::xlab(NULL) + ggplot2::ylab(private$ylab)
+      
+      current_theme <- ggplot2::theme_get()
+      current_theme$text <- element_text(size = 11 * private$text_size)
+      p + current_theme + xlab(private$xlab) + ylab(private$ylab)
     },
 
     #' @description
     #' Generate a Theseus plot for a specified column
     #'
     #' @param column_name The name of the column to visualize.
-    #' @param n integer. Maximum number of top contributing attributes to display.
-    #'   Remaining attributes are aggregated if necessary.
-    #' @param main_item string. The attribute used as the reference for scaling
+    #' @param n integer. Maximum number of top contributing subgroups to display.
+    #'   Remaining subgroups are aggregated if necessary.
+    #' @param main_item string. The subgroup used as the reference for scaling
     #'   the bar heights.
     #' @param bar_max_value numeric. Maximum value for scaling the contribution bars.
-    #' @param levels character vector specifying the display order of attributes.
+    #' @param levels character vector specifying the display order of subgroups
     #' @param continuous list. Configuration for handling continuous variables
     #'   (e.g., number of bins or custom breaks).
     #'
@@ -419,39 +453,32 @@ ShipOfTheseus <- R6::R6Class(
       column_name <- rlang::ensym(column_name) |> rlang::as_string()
 
       labels <- private$labels
-
-      score2 <- private$compute_scores(column_name)[2]
-
-      result <- self$table(!!rlang::sym(column_name), n = n, continuous = continuous) |>
-        select(items = 1L, contrib) |>
-        mutate(contrib = -contrib)
-
-      is_factor <- is.factor(result$items)
-      if (is_factor) {
-        result <- result |> arrange(desc(items))
-      } else {
-        result <- result |> arrange(contrib)
-      }
-      data_size <- private$compute_size(column_name, target = result$items, continuous = continuous)
-
-      if (!is.null(levels)) {
-        levels <- as.character(levels) |> rev()
-        result <- data.frame(items = levels) |> inner_join(result, by = "items")
-      }
-      names <- as.character(result$items)
-      result <- tibble::tibble(items = labels[2], contrib = score2) |>
-        bind_rows(result)|>
+      scores <- private$compute_scores(column_name)
+      score2 <- scores[2]
+      max_score <- max(scores)
+      plot_data <- private$prepare_plot_data(column_name, n, levels, continuous, flip = TRUE)
+      
+      df_contrib <- plot_data$df_contrib
+      data_size <- plot_data$data_size
+      
+      names <- as.character(df_contrib$items)
+      data_waterfall <- tibble::tibble(items = labels[2], contrib = score2) |>
+        bind_rows(df_contrib)|>
         mutate(contrib = round(contrib * 100, digits = private$digits))
 
-      colors <- if_else(result$contrib > 0, "#F8766D", "#00BFC4")
+      colors <- if_else(data_waterfall$contrib > 0, "#F8766D", "#00BFC4")
       colors[1] <- "#00BFC4"
-      p <- waterfalls::waterfall(
-        result, calc_total = TRUE, total_axis_text = labels[1],
+      n_bars <- nrow(data_waterfall) + 1L
+      p <- waterfall(
+        data_waterfall,
+        calc_total = TRUE, total_axis_text = n_bars,
         total_rect_text_color = "black", fill_colours = colors,
         fill_by_sign = FALSE, total_rect_color = "#00BFC4",
         rect_text_size = private$text_size) +
         coord_flip()
-
+      scale_x <- p@scales$get_scales("x")
+      scale_x$labels <- c(data_waterfall$items, labels[1])
+      
       reverse_sign <- function(x) {
         x <- str_replace(x, "\u2212", "-")
         x <- -as.numeric(x)
@@ -470,35 +497,24 @@ ShipOfTheseus <- R6::R6Class(
         }
       }
 
-      if (is.null(main_item) & is.null(bar_max_value)) {
-        data_max <- result |> tail(-1) |> 
-          slice_max(order_by = abs(contrib), n = 1, with_ties = FALSE)
-        max_item <- data_max |> pull(items)
-        max_amount <- data_max |> pull(contrib) |> abs()
-        n_max <- data_size |> filter(items == max_item) |> pull(n) |> max()
-      } else if(!is.null(main_item)) {
-        max_amount <- result |> filter(items == main_item) |> pull(contrib) |> abs()
-        n_max <- data_size |> filter(items == main_item) |> pull(n) |> max()
-      } else if(!is.null(bar_max_value)) {
-        max_amount <- bar_max_value
-        n_max <- data_size |> filter(n == max(n)) |> pull(n) |> max()
-      }
-
-      levels <- c(labels[2], names, labels[1])
-      data_size <- p$data |> select(x) |> distinct() |>
-        left_join(data_size, by = c(x = "items")) |>
+      bar_rate <- private$compute_bar_height(data_waterfall, data_size, max_score, 
+                                             main_item, bar_max_value)
+      
+      data_size <- tibble(x = factor(seq_len(n_bars))) |>
+        mutate(items = c(labels[2], names, labels[1])) |>
+        left_join(data_size, by = "items") |>
         tidyr::replace_na(list(n = 0L)) |>
-        mutate(x = factor(x, levels = levels), type = factor(type, levels = rev(labels))) |>
-        mutate(n = n / n_max * max_amount)
-
+        mutate(type = factor(type, levels = rev(labels))) |>
+        mutate(n = n * bar_rate)
+      
       p <- p +
         geom_col(data = data_size, aes(x, n, fill = type), width = 0.7, position = position_dodge()) +
         scale_fill_manual(values = c("#C77CFF", "#7CAE00"), guide = "none")
       p$layers <- append(head(p$layers, -1), tail(p$layers, 1), 1)
-      theme <- ggplot2::theme_get()
-      p + ggplot2::ggtitle(NULL, subtitle = column_name) +
-        theme(private$text_size * 11) +
-        ggplot2::xlab(NULL) + ggplot2::ylab(private$ylab)
+      
+      current_theme <- ggplot2::theme_get()
+      current_theme$text <- element_text(size = 11 * private$text_size)
+      p + current_theme + xlab(private$xlab) + ylab(private$ylab)
     }
   )
 )